Fix type error on custom calls

+++ kernel/defs.h
@@ -102,11 +102,11 @@ int             either_copyout(int user_dst, uint64 dst, void *src, uint64 len);
 int             either_copyin(void *dst, int user_src, uint64 src, uint64 len);
 void            procdump(void);
 //custom calls in proc.c
-int             getnice(int pid);
-int             setnice(int pid, int value);
+uint64          getnice(int pid);
+uint64          setnice(int pid, int value);
 void            ps(int pid);
-int             meminfo(int pid);
-int             waitpid(int pid);
+uint64          meminfo(int pid);
+uint64          waitpid(int pid);
 
 // swtch.S
 void            swtch(struct context*, struct context*);
